+++ tiledb/libtiledb.pyx
@@ -2910,7 +2910,7 @@ cdef class Array(object):
 
         return tuple(results)
 
-    def consolidate(self, config=None, key=None, timestamp=None):
+    def consolidate(self, config=None, key=None, fragment_uris=None, timestamp=None):
         """
         Consolidates fragments of an array object for increased read performance.
 
@@ -2919,8 +2919,8 @@ cdef class Array(object):
         :param tiledb.Config config: The TileDB Config with consolidation parameters set
         :param key: (default None) encryption key to decrypt an encrypted array
         :type key: str or bytes
-        :param timestamp: (default None) If not None, consolidate the array using the
-            given tuple(int, int) UNIX seconds range (inclusive)
+        :param fragment_uris: (default None) Consolidate the array using a list of fragment file names
+        :param timestamp: (default None) If not None, consolidate the array using the given tuple(int, int) UNIX seconds range (inclusive). This argument will be ignored if `fragment_uris` is passed.
         :type timestamp: tuple (int, int)
         :raises: :py:exc:`tiledb.TileDBError`
 
@@ -2932,7 +2932,8 @@ cdef class Array(object):
         """
         if self.mode == 'r':
             raise TileDBError("cannot consolidate array opened in readonly mode (mode='r')")
-        return consolidate(uri=self.uri, key=key, config=config, ctx=self.ctx, timestamp=timestamp)
+        return consolidate(uri=self.uri, key=key, config=config, ctx=self.ctx, 
+                           fragment_uris=fragment_uris, timestamp=timestamp)
 
     def upgrade_version(self, config=None):
         """
@@ -4454,15 +4455,15 @@ cdef class SparseArrayImpl(Array):
         return dim_values
 
 
-def consolidate(uri, key=None, config=None, ctx=None, timestamp=None):
+def consolidate(uri, key=None, config=None, ctx=None, fragment_uris=None, timestamp=None):
     """Consolidates TileDB array fragments for improved read performance
 
     :param str uri: URI to the TileDB Array
     :param str key: (default None) Key to decrypt array if the array is encrypted
     :param tiledb.Config config: The TileDB Config with consolidation parameters set
     :param tiledb.Ctx ctx: (default None) The TileDB Context
-    :param timestamp: (default None) If not None, consolidate the array using the given
-        tuple(int, int) UNIX seconds range (inclusive)
+    :param fragment_uris: (default None) Consolidate the array using a list of fragment file names
+    :param timestamp: (default None) If not None, consolidate the array using the given tuple(int, int) UNIX seconds range (inclusive). This argument will be ignored if `fragment_uris` is passed.
     :rtype: str or bytes
     :return: path (URI) to the consolidated TileDB Array
     :raises TypeError: cannot convert path to unicode string
@@ -4473,15 +4474,95 @@ def consolidate(uri, key=None, config=None, ctx=None, timestamp=None):
     `"sm.vacuum.timestamp_end"` which takes in a time in UNIX seconds. If both
     are set then this function's `timestamp` argument will be used.
 
+    **Example:**
+
+    >>> import tiledb, tempfile, numpy as np, os
+    >>> path = tempfile.mkdtemp()
+
+    >>> with tiledb.from_numpy(path, np.zeros(4), timestamp=1) as A:
+    ...     pass
+    >>> with tiledb.open(path, 'w', timestamp=2) as A:
+    ...     A[:] = np.ones(4, dtype=np.int64)
+    >>> with tiledb.open(path, 'w', timestamp=3) as A:
+    ...     A[:] = np.ones(4, dtype=np.int64)
+    >>> with tiledb.open(path, 'w', timestamp=4) as A:
+    ...     A[:] = np.ones(4, dtype=np.int64)
+    >>> len(tiledb.array_fragments(path))
+    4
+
+    >>> fragment_names = [
+    ...     os.path.basename(f) for f in tiledb.array_fragments(path).uri
+    ... ]
+    >>> array_uri = tiledb.consolidate(
+    ...    path, fragment_uris=[fragment_names[1], fragment_names[3]]
+    ... )
+    >>> len(tiledb.array_fragments(path))
+    3
+
     """
     if not ctx:
         ctx = default_ctx()
 
+    if fragment_uris is not None:
+        if timestamp is not None:
+            warnings.warn(
+                "The `timestamp` argument will be ignored and only fragments "
+                "passed to `fragment_uris` will be consolidate",
+                DeprecationWarning,
+            )
+        return _consolidate_uris(
+            uri=uri, key=key, config=config, ctx=ctx, fragment_uris=fragment_uris)
+    else:
+        return _consolidate_timestamp(
+            uri=uri, key=key, config=config, ctx=ctx, timestamp=timestamp)
+
+def _consolidate_uris(uri, key=None, config=None, ctx=None, fragment_uris=None):
+    cdef int rc = TILEDB_OK
+
     cdef tiledb_ctx_t* ctx_ptr = safe_ctx_ptr(ctx)
 
-    if timestamp:
+    if config is None:
+        config = ctx.config()
+
+    cdef tiledb_config_t* config_ptr = NULL
+    if config is not None:
+        config_ptr = <tiledb_config_t*>PyCapsule_GetPointer(
+            config.__capsule__(), "config")
+    cdef bytes buri = unicode_path(uri)
+    cdef const char* array_uri_ptr = PyBytes_AS_STRING(buri)
+
+    cdef const char **fragment_uri_buf = <const char **>malloc(
+        len(fragment_uris) * sizeof(char *))
+
+    for i, frag_uri in enumerate(fragment_uris):
+        fragment_uri_buf[i] = PyUnicode_AsUTF8(frag_uri)
+
+    if key is not None:
+        config["sm.encryption_key"] = key
+
+    rc = tiledb_array_consolidate_fragments(
+        ctx_ptr, array_uri_ptr, fragment_uri_buf, len(fragment_uris), config_ptr)
+    if rc != TILEDB_OK:
+        _raise_ctx_err(ctx_ptr, rc)
+    
+    free(fragment_uri_buf)
+
+    return uri
+
+def _consolidate_timestamp(uri, key=None, config=None, ctx=None, timestamp=None):
+    cdef int rc = TILEDB_OK
+
+    cdef tiledb_ctx_t* ctx_ptr = safe_ctx_ptr(ctx)
+
+    if timestamp is not None:
+        warnings.warn(
+            "The `timestamp` argument is deprecated; pass a list of "
+            "fragment URIs to consolidate with `fragment_uris`",
+            DeprecationWarning,
+        )
+
         if config is None:
-            config = Config()
+            config = ctx.config()
 
         if not isinstance(timestamp, tuple) and len(timestamp) != 2:
             raise TypeError("'timestamp' argument expects tuple(start: int, end: int)")
@@ -4496,7 +4577,8 @@ def consolidate(uri, key=None, config=None, ctx=None, timestamp=None):
         config_ptr = <tiledb_config_t*>PyCapsule_GetPointer(
             config.__capsule__(), "config")
     cdef bytes buri = unicode_path(uri)
-    cdef const char* uri_ptr = PyBytes_AS_STRING(buri)
+    cdef const char* array_uri_ptr = PyBytes_AS_STRING(buri)
+
     # encryption key
     cdef:
         bytes bkey
@@ -4514,9 +4596,9 @@ def consolidate(uri, key=None, config=None, ctx=None, timestamp=None):
         #TODO: unsafe cast here ssize_t -> uint64_t
         key_len = <unsigned int> PyBytes_GET_SIZE(bkey)
 
-    cdef int rc = TILEDB_OK
     with nogil:
-        rc = tiledb_array_consolidate_with_key(ctx_ptr, uri_ptr, key_type, key_ptr, key_len, config_ptr)
+        rc = tiledb_array_consolidate_with_key(
+            ctx_ptr, array_uri_ptr, key_type, key_ptr, key_len, config_ptr)
     if rc != TILEDB_OK:
         _raise_ctx_err(ctx_ptr, rc)
     return uri